Hash all assets to make them immutable.
Create a SHA-1 hash of each file's contents, then append part of the
hash to the filename. This ensures the path will change when the
contents do, so we can tell the browser to cache them forever.

Also refactor page handling to use the same async iterable helper
as the assets.